refactor: move core API functions to core-api.ts for better organization
